requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-allauth==0.45.0
django-crispy-forms==1.13.0
idna==3.2
oauthlib==3.1.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Pillow==8.3.2
gunicorn==20.1.0
psycopg2==2.9.1
